Reimagine hiera config
If values are defined as :key: 'value' in hiera the keys actually
become symbols when they reach templates.  This has been changed so
the keys are strings regardless of whether they come from hiera or
the DSL to avoid problems in future.  Generation of the file has
been delegated to a rudimentary yaml dump function which handles the
creation of puppet specific yaml transparently